Se realizo la integracion del pid con el motor udp, se agrego algrafico el torque

+++ cpp/pid/PID.h
@@ -3,26 +3,26 @@
 class PID {
     public:
         PID (float kp, float ki, float kd);
-        void setDesiredSpeed (float);
-        float getDesiredSpeed ();
-        float getActualSpeed ();
-        void setActualSpeed (float);
+        void setDesiredSpeed (int);
+        int getDesiredSpeed ();
+        int getActualSpeed ();
+        void setActualSpeed (int);
         void setKp (float);
         void setKi (float);
         void setKd (float);
         void setDt (float);
-        float getTorque ();
-        float calculateTorque ();
+        int getTorque ();
+        int calculateTorque ();
         
     private:
-        float desiredSpeed;
-        float actualSpeed;
+        int desiredSpeed;
+        int actualSpeed;
         float kp;
         float ki;
         float kd;
         float dt;
         float lastError;
         float acumulatedError;
-        float torque;
+        int torque;
 };
 #endif
